refactor(plugin-tower): move EndpointController to package plugin/tower/pkg/controller/endpoint
