+++ plugin/tower/pkg/controller/endpoint/controller_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller_test
+package endpoint_test
 
 import (
 	"context"
@@ -27,7 +27,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/rand"
 
-	"github.com/smartxworks/lynx/plugin/tower/pkg/controller"
+	controller "github.com/smartxworks/lynx/plugin/tower/pkg/controller/endpoint"
 	"github.com/smartxworks/lynx/plugin/tower/pkg/schema"
 )
